Add option to force usage of docker expose flag
Until now it was possible to expose meteor container port only when
using let's encrypt or proxy. In other cases `start.sh` was binding to
ports on host. If you used external loadbalancer you were forced to
track which ports are bound already. This it total PITA when you have
a few apps.

This change adds option to force use of docker's expose option.